# release-23.1: rttanalysis, catalog: add benchmark for scanning comments, database-aware scans for comments